Lower-case boolean literals 'True' and 'False'
In https://github.com/masak/007/pull/451 the three literals `None`,
`True`, and `False` changed capitalization to `none`, `true`, and
`false`, respectively. This commit tracks that.

I took the chance to also replace a few `== False` comparisons with
boolean negations. There's a style rule in there somewhere -- not
to compare directly against booleans in conditionals, instead
falling back on truthiness semantics. And yes, this is _despite_
the fact that writing it out as the direct comparison is actually
more exact -- "vagueness is a virtue" in this case, it seems --
and also despite the fact in these two instances, the variables
compared are boolean variables, could be typed as such, and then
truthiness would come down to direct comparison; that is, there's
no semantic difference _at all_ between the two styles.  Still, I
find I instinctively want to change `== True` to just the lhs
boolean and `== False` to just its negation... direct comparisons
with booleans look redundant to me, on a fairly instinctual level.

I will try to keep this rule in mind as I pen the actual Style
Guide™, coming soon to a repository near you!